Rewrite Mainable and company
This rewrite lets us use the Mainable class without the backend token on
the diagram. It also allows using multiple backends with a single
executable. This is done by adding another parameter to the class, one
for the token used to render and the other is thing we are rendering.

It should be more versitile, allowing custom mainWith-like functions
since mainWith is no longer part of the class.

I'm still unsure about the names I've chosen.